+++ lib/kandr-rails-extensions/time.rb
@@ -1,4 +1,9 @@
 module KandrRailsExtensions
+  module TimeEpochTime
+    alias_method :to_epoch, :to_i
+    alias_method :to_posix, :to_i
+    alias_method :to_unix, :to_i
+  end
 
   module TimeMilliseconds
     def to_ms
@@ -8,5 +13,6 @@ module KandrRailsExtensions
 end
 
 class Time
+  include KandrRailsExtensions::TimeEpochTime
   include KandrRailsExtensions::TimeMilliseconds
 end